Implement changes to default config file persistency
- Disable "Local PC" gametype

BwHeadless.java
- Implement "boolean writeDefaultConfigFile()" to call
after a variable is set.

ConfigFile.java
- Change implementation so that the configuration file changes
get written to disk when a variable is changed.

Notes:
- Should now be able to use "BwHeadless" set functions such as
"setBotClient(File file)" and "setBotDll(File file)" to
store their values in memory and on disk. When a variable is set,
it gets written to the default configuration file:
"BwHeadless.DEFAULT_CFG_FILE".

+++ src/droplauncher/tools/ProcessPipe.java
@@ -93,7 +93,7 @@ public final class ProcessPipe {
 
       //      this.is = this.process.getInputStream();
       //      this.br = new BufferedReader(new InputStreamReader(this.is));
-      StreamGobbler gobblerStdout = new StreamGobbler(this.is);
+      StreamGobbler gobblerStdout = new StreamGobbler(this.process.getInputStream());
       StreamGobbler gobblerStderr = new StreamGobbler(this.process.getErrorStream());
       gobblerStdout.start();
       gobblerStderr.start();
@@ -119,15 +119,17 @@ public final class ProcessPipe {
    */
   public boolean close() {
     try {
-      if (this.process != null && this.process.isAlive()) {
+      if (this.br != null && this.is != null
+          && this.bw != null && this.os != null
+          && this.process != null && this.process.isAlive()) {
+        this.process.destroyForcibly();
         this.br.close();
         this.is.close();
         this.bw.close();
         this.os.close();
-        this.process.destroyForcibly();
       }
       return true;
-    } catch (IOException ex) {
+    } catch (Exception ex) {
       LOGGER.error(ex.getMessage(), ex);
     }
     return false;
